width and height changed to int

+++ src/edu/hm/fuberg/se2/android/robotz/control/PillChecker.java
@@ -64,8 +64,8 @@ class PillChecker {
 
 		while (noFreeSlot) {
 
-			final int xCoord = random.nextInt((int) robotzData.getWidth());
-			final int yCoord = random.nextInt((int) robotzData.getHeight());
+			final int xCoord = random.nextInt(robotzData.getWidth());
+			final int yCoord = random.nextInt(robotzData.getHeight());
 
 			if (!invinciblePillOnItem(xCoord, yCoord)) {
